pyproject.toml
[project]
name = "fastapi-jwt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.120.1",
    "passlib[bcrypt]>=1.7.4",
    "pyjwt>=2.10.1",
]
